fix saving of parameters
Fix saving of parameters.

Some of the crashes observed seem to be caused by an
overload of the SERCOS device if too many parameters
are retrieved in short time. It seems that throttling
the retrieval helps here.

WIP: The save-all dialog with the progressbar
is somehow present but it does not show up.